requirements.txt
Flask==1.1.1
requests2==2.16.0
pytest==5.1.0
